+++ docker-gen-proto.sh
@@ -31,5 +31,5 @@ gen_proto_go checkoutservice
 #gen_proto_js payment
 gen_proto_go productcatalogservice
 #gen_proto_php quote
-gen_proto_python recommendationservice
+gen_proto_python recommendation
 #gen_proto_rust shipping