+++ ide-gen-proto.sh
@@ -76,5 +76,5 @@ gen_proto_ts react-native-app
 gen_proto_js payment
 gen_proto_go productcatalogservice
 # gen_proto_php quote
-gen_proto_python recommendationservice
+gen_proto_python recommendation
 gen_proto_rust shipping